Leverage retworkx for computing number of shortest paths
In retworkx 0.10.0 a new function, num_shortest_paths_unweighted(), was
added that performs the same function of GraphUtils.num_shortest_path()
except it is significantly faster. This commit removes the graph_utils
module (as it only contained the one function) and replaces the one use
of GraphUtils.num_shortest_path() with
retworkx.num_shortest_paths_unweighted(). The minimum required retworkx
version is also increased to 0.10.0.

Fixes #26