Generalized circ.py for various types
Updated circ.py so now user has the ability to create a TopologicalCircuit with multiple types of logical qubits, not just one singular type. This is done by constructing a treg object with a list of ctypes, e.g - [REPETITION, XXZZ, XZZX], and a list of params, e.g - [{"d": 3}, {"d": 3}], both with default options. Also got rid of add_tqubit, as it was redundant now. add_tqubits now takes arguments as well, same with cx.